created route to policy specific articles page
