+++ server.js
@@ -3,6 +3,7 @@ const routes = require("./controllers");
 const sequelize = require("./config/connection.js");
 const path = require("path");
 const exphbs = require("express-handlebars");
+//const helpers = require("./utils/helpers");
 const hbs = exphbs.create({});
 
 const app = express();
