+++ views/policy.handlebars
@@ -1,0 +1,11 @@
+<ul class="article-list list-group-flsuh m-2">
+  {{#each policy.articles}}
+    <li class="article-list-itm list-group item card border border-6 mx-auto">
+      <article class="article card-body">
+        <div class="title">
+          {{title}}
+        </div>
+        <div class="author"></div>
+      </article></li>
+  {{/each}}
+</ul>
